# docs/xaucloud-gpt-actions.openapi.yaml
openapi: 3.1.0
info:
  title: XauCloud Admin Marketing Actions
  version: 2.0.0
  description: >-
    Private administrative email and controlled marketing actions for XauCloud.
    Email operations use the existing XauCloud drafts, renderer, audience resolver,
    delivery provider, and history. Marketing operations expose only approved facts
    and predefined campaign surfaces. They never expose SMTP, database, or hosting
    credentials.
servers:
  - url: https://xaucloud.io/api
security:
  - actionBearer: []
paths:
...
  /admin/actions/email/drafts:
    post:
      operationId: createEmailDraft
      summary: Create a structured XauCloud email draft
      description: >-
        Create a draft in the same draft library used by the XauCloud Admin
        composer. Use structured blocks; do not supply arbitrary full HTML.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDraftRequest"
      responses:
        "200":
          description: Created draft.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Draft" }
  /admin/actions/email/drafts/{id}:
    get:
      operationId: getEmailDraft
      summary: Retrieve a saved email draft
      parameters:
        - $ref: "#/components/parameters/DraftId"
      responses:
        "200":
          description: Current authoritative draft.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Draft" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/actions/email/drafts/{id}/preview:
    post:
      operationId: previewEmailDraft
      summary: Render a draft with the authoritative XauCloud renderer
      description: Use this to show the owner a concise plain-text representation and preview metadata.
      parameters:
        - $ref: "#/components/parameters/DraftId"
      responses:
        "200":
          description: Sanitized renderer output and metadata.
          content:
            application/json:
              schema:
                type: object
                properties:
                  draft_id: { type: string }
                  subject: { type: string }
                  audience: { $ref: "#/components/schemas/Audience" }
                  sender: { type: string }
                  preview_text: { type: string }
                  text: { type: string }
                  renderer: { type: string }
  /admin/actions/email/drafts/{id}/test:
    post:
      operationId: sendTestEmail
      summary: Send one test email only
      description: Sends the exact draft through the existing XauCloud provider to one explicitly validated test address. It never sends to the draft audience.
      parameters:
        - $ref: "#/components/parameters/DraftId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to]
              properties:
                to: { type: string, format: email, maxLength: 320 }
      responses:
        "200":
          description: Test provider acceptance result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sent: { type: boolean }
                  test_only: { type: boolean, const: true }
                  to: { type: string, format: email }
                  draft_id: { type: string }
  /admin/actions/email/drafts/{id}/prepare-send:
    post:
      operationId: prepareEmailBroadcast
      summary: Validate and prepare a campaign without sending it
      description: >-
        Resolve the current audience, validate the draft and renderer, and issue
        a short-lived single-use confirmation token. This operation never sends
        a production email. Present subject, audience, recipient_count, sender,
        warnings, and expiration to the owner before asking for explicit approval.
      parameters:
        - $ref: "#/components/parameters/DraftId"
      responses:
        "200":
          description: Prepared campaign and confirmation token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PreparedCampaign" }
  /admin/actions/email/drafts/{id}/send:
    post:
      operationId: sendEmailBroadcast
      summary: Send a previously prepared campaign to real XauCloud users
      description: >-
        This action sends an email to real XauCloud users and must only be called
        after the user explicitly confirms the prepared campaign. Never call it
        from an inferred or ambiguous instruction. Use exactly the confirmation
        token returned by prepareEmailBroadcast and a stable unique idempotency
        key. Do not place subject, content, or audience in this request.
      parameters:
        - $ref: "#/components/parameters/DraftId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [confirmation_token, idempotency_key]
              properties:
                confirmation_token:
                  type: string
                  description: Short-lived single-use token returned by prepareEmailBroadcast.
                idempotency_key:
                  type: string
                  minLength: 8
                  maxLength: 128
                  pattern: "^[A-Za-z0-9._:-]+$"
                  description: Stable unique value reused unchanged if this exact send request is retried.
      responses:
        "200":
          description: Accepted or completed send result. A retry returns duplicate=true and the original result.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BroadcastStatus" }
        "409":
          description: Draft, audience, recipients, or confirmation state changed; prepare again and obtain fresh approval.
  /admin/actions/email/broadcasts/{id}:
    get:
      operationId: getBroadcastStatus
      summary: Get the backend-confirmed status of a GPT-initiated broadcast
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, maxLength: 100 }
      responses:
        "200":
          description: Current persisted result from the normal XauCloud email history.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BroadcastStatus" }
...
components:
  securitySchemes:
    actionBearer:
      type: http
      scheme: bearer
      description: Dedicated XAUCLOUD_GPT_ACTION_SECRET. This is not an SMTP, mailbox, database, or hosting credential.
  parameters:
    DraftId:
      name: id
      in: path
      required: true
      schema: { type: string, maxLength: 100 }
...
  responses:
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Audience:
      type: string
      enum: [single, all_users, customers, active_license, inactive_license, selected]
    EmailTheme:
      type: object
      additionalProperties: false
      required: [width, background, contentBackground, accent, radius, spacing]
      properties:
        width: { type: integer, enum: [600, 640, 680] }
        background: { type: string, enum: ["#08080A", "#111114", "#19191D", "#FFFFFF", "#F5F0E1", "#D6B35A", "#B88716", "#25252A", "#8E8E97"] }
        contentBackground: { type: string, enum: ["#08080A", "#111114", "#19191D", "#FFFFFF", "#F5F0E1", "#D6B35A", "#B88716", "#25252A", "#8E8E97"] }
        accent: { type: string, enum: ["#D6B35A", "#B88716"] }
        radius: { type: integer, enum: [0, 6, 10, 14] }
        spacing: { type: string, enum: [compact, normal, spacious] }
    EmailItem:
      type: object
      additionalProperties: false
      required: [label]
      properties:
        label: { type: string, maxLength: 120 }
        value: { type: string, maxLength: 120 }
        text: { type: string, maxLength: 1000 }
        title: { type: string, maxLength: 300 }
    EmailColumn:
      type: object
      additionalProperties: false
      properties:
        title: { type: string, maxLength: 300 }
        html: { type: string, maxLength: 20000, description: Sanitized rich-text fragment only. }
    EmailBlock:
      type: object
      additionalProperties: false
      required: [id, type]
      properties:
        id: { type: string, minLength: 1, maxLength: 100 }
        type: { type: string, enum: [hero, announcement, text, heading, button, image, divider, spacer, callout, metrics, feature, steps, columns, section, risk, footer] }
        html: { type: string, maxLength: 100000, description: Sanitized rich-text fragment; scripts and unsafe markup are removed by XauCloud. }
        text: { type: string, maxLength: 10000 }
        title: { type: string, maxLength: 500 }
        subtitle: { type: string, maxLength: 2000 }
        badge: { type: string, maxLength: 80 }
        level: { type: integer, minimum: 1, maximum: 3 }
        url: { type: string, maxLength: 2000, description: HTTP or HTTPS only; unsafe schemes are discarded. }
        alt: { type: string, maxLength: 500 }
        link: { type: string, maxLength: 2000 }
        width: { type: integer, minimum: 80, maximum: 640 }
        align: { type: string, enum: [left, center, right] }
        style: { type: string, enum: [gold, outline, dark] }
        fullWidth: { type: boolean }
        tone: { type: string, enum: [neutral, gold, warning] }
        height: { type: integer, minimum: 8, maximum: 120 }
        items: { type: array, maxItems: 8, items: { $ref: "#/components/schemas/EmailItem" } }
        columns: { type: array, minItems: 2, maxItems: 2, items: { $ref: "#/components/schemas/EmailColumn" } }
        background: { type: string, enum: ["#08080A", "#111114", "#19191D", "#FFFFFF", "#F5F0E1", "#D6B35A", "#B88716", "#25252A", "#8E8E97"] }
        padding: { type: string, enum: [compact, normal, spacious] }
    EmailDocument:
      type: object
      additionalProperties: false
      required: [version, theme, blocks]
      properties:
        version: { type: integer, const: 1 }
        theme: { $ref: "#/components/schemas/EmailTheme" }
        blocks:
          type: array
          minItems: 1
          maxItems: 100
          items: { $ref: "#/components/schemas/EmailBlock" }
    CreateDraftRequest:
      type: object
      additionalProperties: false
      required: [subject, audience, document]
      properties:
        campaign_id: { type: string, maxLength: 100, description: Optional parent marketing campaign id. }
        title: { type: string, maxLength: 160 }
        subject: { type: string, minLength: 1, maxLength: 300 }
        preheader: { type: string, maxLength: 300 }
        audience: { $ref: "#/components/schemas/Audience" }
        to: { type: string, format: email, maxLength: 320 }
        selected_recipients: { type: array, maxItems: 250, items: { type: string, format: email } }
        document: { $ref: "#/components/schemas/EmailDocument" }
    Draft:
      allOf:
        - $ref: "#/components/schemas/CreateDraftRequest"
        - type: object
          properties:
            id: { type: string }
            preview_text: { type: string }
            source: { type: string }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    PreparedCampaign:
      type: object
      properties:
        draft_id: { type: string }
        subject: { type: string }
        audience: { $ref: "#/components/schemas/Audience" }
        recipient_count: { type: integer }
        sender: { type: string }
        warnings: { type: array, items: { type: string } }
        confirmation_token: { type: string }
        confirmation_expiration: { type: string, format: date-time }
    BroadcastStatus:
      type: object
      properties:
        broadcast_id: { type: string }
        draft_id: { type: string }
        subject: { type: string }
        audience: { $ref: "#/components/schemas/Audience" }
        recipient_count: { type: integer }
        sender: { type: string }
        status: { type: string, enum: [sending, sent, partial, failed] }
        sent: { type: integer }
        failed: { type: integer }
        source: { type: string, const: chatgpt_action }
        duplicate: { type: boolean }
        error: { type: string }
    Error:
      type: object
      properties:
        detail: { type: string }